Add Student controller that extends User controller
