+++ app/Config/Routes.php
@@ -2,6 +2,7 @@
 
 use CodeIgniter\Router\RouteCollection;
 use App\Controllers\User;
+use App\Controllers\Student;
 
 /**
  * @var RouteCollection $routes
@@ -14,7 +15,7 @@ $routes->view('/signup', 'register_student');
 
 // handle login and signup requests
 $routes->post('/login', [User::class, 'login']);
-$routes->post('/signup', [User::class, 'signup']);
+$routes->post('/signup', [Student::class, 'signup']);
 
 // handle logout request
 $routes->get('/logout', [User::class, 'logout']);
